Switch MySQL plugin to use 'gauge' value type
The 'counter' value type calculates a derivative of the metric, assuming that
the value only increases (and assuming wraparound if the value decreases). Some
of these metrics are more interesting as absolute values (e.g.
threads_connected), and many will decrease as well as increase.

We therefore switch to 'gauge', which stores the value with no modification,
allowing the user to look at the derivative as a post-processing step if using
e.g. Graphite.